Add script to build docker image and deploy to k8s
This commit adds script to build the docker image in order to deploy
to k8s through TravisCI. The image is tagged with the git SHA so that
k8s recognized the difference in the image and do the necessary
deployments.

Also add a quick fix to CLOUDSDK_CORE_DISABLE_PROMPTS env to disable
GCloud CLI to prompt anything.